# docs/swagger.yaml
definitions:
...
  resource.ModuleVO:
    properties:
      category:
        type: string
      code:
        type: string
      color:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      icon:
        type: string
      id:
        type: string
      is_visible:
        type: string
      name:
        type: string
      sort_code:
        type: integer
      status:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  resource.ResourceVO:
    properties:
      category:
        type: string
      children:
        items:
          $ref: '#/definitions/resource.ResourceVO'
        type: array
      code:
        type: string
      color:
        type: string
      component_path:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      external_url:
        type: string
      extra:
        type: string
      icon:
        type: string
      id:
        type: string
      is_affix:
        type: string
      is_breadcrumb:
        type: string
      is_cache:
        type: string
      is_visible:
        type: string
      name:
        type: string
      parent_id:
        type: string
      redirect_path:
        type: string
      route_path:
        type: string
      sort_code:
        type: integer
      status:
        type: string
      type:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
  utils.IdsParam:
    properties:
      ids:
        items:
          type: string
        type: array
    type: object
info:
  contact:
    name: Charlie
    url: https://github.com/jiangbyte
  description: Hei Gin 后台管理系统 API 文档
  license:
    name: MIT
    url: https://github.com/jiangbyte/hei-gin/blob/main/LICENSE
  termsOfService: https://github.com/jiangbyte/hei-gin
  title: Hei Gin API
  version: 1.0.0
paths:
...
  /api/v1/sys/module/create:
    post:
      consumes:
      - application/json
      description: 访问 /api/v1/sys/module/create，资源管理创建
      parameters:
      - description: 请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/resource.ModuleVO'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: 资源管理创建
      tags:
      - 资源管理
  /api/v1/sys/module/detail:
    get:
      consumes:
      - application/json
      description: 访问 /api/v1/sys/module/detail，资源管理详情查询
      parameters:
      - description: id
        in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: 资源管理详情查询
      tags:
      - 资源管理
  /api/v1/sys/module/modify:
    post:
      consumes:
      - application/json
      description: 访问 /api/v1/sys/module/modify，资源管理修改
      parameters:
      - description: 请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/resource.ModuleVO'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: 资源管理修改
      tags:
      - 资源管理
  /api/v1/sys/module/page:
    get:
      consumes:
      - application/json
      description: 访问 /api/v1/sys/module/page，资源管理分页查询
      parameters:
      - in: query
        name: current
        type: integer
      - in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: 资源管理分页查询
      tags:
      - 资源管理
  /api/v1/sys/module/remove:
    post:
      consumes:
      - application/json
      description: 访问 /api/v1/sys/module/remove，资源管理删除
      parameters:
      - description: 请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/utils.IdsParam'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: 资源管理删除
      tags:
      - 资源管理
...
  /api/v1/sys/resource/create:
    post:
      consumes:
      - application/json
      description: 访问 /api/v1/sys/resource/create，资源管理创建
      parameters:
      - description: 请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/resource.ResourceVO'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: 资源管理创建
      tags:
      - 资源管理
  /api/v1/sys/resource/detail:
    get:
      consumes:
      - application/json
      description: 访问 /api/v1/sys/resource/detail，资源管理详情查询
      parameters:
      - description: id
        in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: 资源管理详情查询
      tags:
      - 资源管理
  /api/v1/sys/resource/modify:
    post:
      consumes:
      - application/json
      description: 访问 /api/v1/sys/resource/modify，资源管理修改
      parameters:
      - description: 请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/resource.ResourceVO'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: 资源管理修改
      tags:
      - 资源管理
  /api/v1/sys/resource/page:
    get:
      consumes:
      - application/json
      description: 访问 /api/v1/sys/resource/page，资源管理分页查询
      parameters:
      - in: query
        name: current
        type: integer
      - in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: 资源管理分页查询
      tags:
      - 资源管理
  /api/v1/sys/resource/remove:
    post:
      consumes:
      - application/json
      description: 访问 /api/v1/sys/resource/remove，资源管理删除
      parameters:
      - description: 请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/utils.IdsParam'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: 资源管理删除
      tags:
      - 资源管理
...
securityDefinitions:
  BearerAuth:
    description: Bearer token authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"